move ephemeris downloader to utils
