--- docs/source/reference/parsers/ephemeris.rst
@@ -1,7 +1,0 @@
-ephemeris module
-================
-
-.. automodule:: ephemeris
-   :members:
-   :undoc-members:
-   :show-inheritance:
