+++ docs/source/reference/parsers/modules.rst
@@ -5,7 +5,6 @@ parsers
    :maxdepth: 4
 
    android
-   ephemeris
    navdata
    nmea
    precise_ephemerides
