+++ docs/source/reference/test_parsers/modules.rst
@@ -5,7 +5,6 @@ parsers
    :maxdepth: 4
 
    test_android
-   test_ephemeris
    test_navdata
    test_nmea
    test_precise_ephemerides
