--- docs/source/reference/test_parsers/test_ephemeris.rst
@@ -1,7 +1,0 @@
-test\_ephemeris module
-======================
-
-.. automodule:: test_ephemeris
-   :members:
-   :undoc-members:
-   :show-inheritance:
