+++ docs/source/reference/test_utils/test_ephemeris_downloader.rst
@@ -1,0 +1,7 @@
+test\_ephemeris\_downloader module
+==================================
+
+.. automodule:: test_ephemeris_downloader
+   :members:
+   :undoc-members:
+   :show-inheritance:
