+++ docs/source/reference/utils/ephemeris_downloader.rst
@@ -1,0 +1,7 @@
+ephemeris\_downloader module
+============================
+
+.. automodule:: ephemeris_downloader
+   :members:
+   :undoc-members:
+   :show-inheritance:
