+++ tests/utils/conftest.py
@@ -14,7 +14,7 @@ import numpy as np
 from gnss_lib_py.utils.time_conversions import gps_millis_to_datetime
 from gnss_lib_py.parsers.navdata import NavData
 from gnss_lib_py.parsers.android import AndroidDerived2022, AndroidGroundTruth2022
-from gnss_lib_py.parsers.ephemeris import EphemerisManager
+from gnss_lib_py.utils.ephemeris_downloader import EphemerisDownloader
 
 @pytest.fixture(name="root_path")
 def fixture_root_path():
@@ -287,7 +287,7 @@ def fixture_all_gps_ephem(ephemeris_path, start_time, all_gps_sats):
         NavData instance containing ephemeris parameters for all GPS
         satellites at the start time for measurement reception.
     """
-    ephem_man_nav = EphemerisManager(ephemeris_path)
+    ephem_man_nav = EphemerisDownloader(ephemeris_path)
     ephem = ephem_man_nav.get_ephemeris(start_time, all_gps_sats)
     return ephem
